requirements.txt
streamlit==1.40.1
pytesseract==0.3.10
opencv-python==4.12.0.88
numpy==1.24.4
spacy==3.8.2
scikit-learn==1.4.2
xgboost==2.0.3
Pillow==10.4.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz
